requirements.txt
git+https://github.com/adafruit/Adafruit_Python_DHT#egg=Adafruit_DHT # --global-option='--force-pi2'
RPi.GPIO>=0.6.3
RPLCD>=1.0.0
